\name{onewaytests-package}
\alias{onewaytests-package}
\alias{onewaytests-package}
\docType{package}
\title{
One-Way Independent Groups Design
}
\description{
Two or more samples in one-way independent groups design are compared via different tests. 
}
\details{
\tabular{ll}{
Package: \tab onewaytests\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2016-07-29\cr
License: \tab GPL (>=2)\cr
}
}