\name{onewaytests-package}
\alias{onewaytests-package}
\alias{onewaytests-package}
\docType{package}
\title{
One-Way Tests in Independent Groups Designs
}
\description{
Performs one-way tests in independent groups designs, pairwise comparisons, graphical approaches, and assess variance homogeneity and normality of data in each group via tests and plots.
}
\details{
\tabular{ll}{
Package: \tab onewaytests\cr
Type: \tab Package\cr
License: \tab GPL (>=2)\cr
}
}