\name{onewaytests-package}
\alias{onewaytests-package}
\alias{onewaytests-package}
\docType{package}
\title{
One-Way Tests in Independent Groups Designs
}
\description{
Performs one-way tests in independent groups designs, pairwise comparisons, graphical approaches, and assess variance homogeneity and normality of each group via tests and plots.
}
\details{
\tabular{ll}{
Package: \tab onewaytests\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2017-04-14\cr
License: \tab GPL (>=2)\cr
}
}