\name{onewaytests-package}
\alias{onewaytests-package}
\alias{onewaytests-package}
\docType{package}
\title{
One-Way Tests for Independent Groups Designs
}
\description{
Performs one-way tests for independent groups designs, graphical approaches, and assess variance homogeneity and normality of each group via tests and plots. 
}
\details{
\tabular{ll}{
Package: \tab onewaytests\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2017-03-16\cr
License: \tab GPL (>=2)\cr
}
}